Move to new temporal architecture
