+++ front/temporal/labs/worker.ts
@@ -1,10 +1,10 @@
 import type { Context } from "@temporalio/activity";
 import { Worker } from "@temporalio/worker";
 
-import * as activities from "@app/lib/labs/temporal/activities";
 import { getTemporalWorkerConnection } from "@app/lib/temporal";
 import { ActivityInboundLogInterceptor } from "@app/lib/temporal_monitoring";
 import logger from "@app/logger/logger";
+import * as activities from "@app/temporal/labs/activities";
 
 import { QUEUE_NAME } from "./config";
